W7/rest-in-peace/requirement.txt
PyJWT
fastapi
pydantic
python-dotenv
uvicorn
